Image Classification Project
1. Problem

As the second-largest provider of carbohydrates in Africa, cassava is a key food security crop grown by smallholder farmers because it can withstand harsh conditions. At least 80% of household farms in Sub-Saharan Africa grow this starchy root, but viral diseases are major sources of poor yields. With the help of data science, it may be possible to identify common diseases so they can be treated.

The task at hand here is to classify each cassava image into four disease categories or a fifth category indicating a healthy leaf. 

2. Data

In this competition, we introduce a dataset of 21,367 labeled images collected during a regular survey in Uganda. Most images were crowdsourced from farmers taking photos of their gardens, and annotated by experts at the National Crops Resources Research Institute (NaCRRI) in collaboration with the AI lab at Makerere University, Kampala. This is in a format that most realistically represents what farmers would need to diagnose in real life.


 * Files
 
  * [train/test]_images the image files. The full set of test images will only be available to your notebook when it is submitted for scoring. Expect to see roughly 15,000 images in the test set.

  * train.csv

    * image_id the image file name.
    * label the ID code for the disease.

  * sample_submission.csv A properly formatted sample submission, given the disclosed test set content.

    * image_id the image file name.
    * label the predicted ID code for the disease.
    
  * [train/test]_tfrecords the image files in tfrecord format.
  
  * label_num_to_disease_map.json , The mapping between each disease code and the real disease name.

3. Evaluation

Submissions will be evaluated based on their categorization accuracy.